+++ src/main/java/org/prgrms/springbootbasic/type/Menu.java
@@ -1,4 +1,4 @@
-package org.prgrms.springbootbasic;
+package org.prgrms.springbootbasic.type;
 
 import java.util.Arrays;
