+++ src/main/java/org/prgrms/springbootbasic/type/TypeOption.java
@@ -1,4 +1,4 @@
-package org.prgrms.springbootbasic;
+package org.prgrms.springbootbasic.type;
 
 public enum TypeOption {
     FIXED("1", "fixedAmountVoucherFactory"), PERCENT("2", "PercentAmountVoucherFactory");
